Copy all the theme overrides to this repository
Instead of maitnaining a dedicated branch in geolexica-server, simply
keep overrides in this repository.

All the modified files are copied from the iec-electropedia branch
in geolexica/geolexica-server repository in their shape as of
https://github.com/geolexica/geolexica-server/commit/65ab5eaf2d2d913.

It is expected that these overrides will be subsequently removed
in favour of _config.yml settings as the Geolexica frameworks gets
more and more customizable.