Make this site working with the most recent Jekyll-Geolexica (#25)
This pull request brings the compatibility with the latest of Jekyll-Geolexica unpatched version. Some commits are purposely keeping known bugs alive — the idea was to have as little diff as possible with the previous state of site.

- Update Jekyll-Geolexica to the latest version.
- Copy all the theme overrides to this repository — they used to be scattered, some were here, some were patches to jekyll-geolexica.
- Update Jekyll-Geolexica configuration so that the site is the same as before.

------

* Copy all the theme overrides to this repository

Instead of maitnaining a dedicated branch in geolexica-server, simply
keep overrides in this repository.

All the modified files are copied from the iec-electropedia branch
in geolexica/geolexica-server repository in their shape as of
https://github.com/geolexica/geolexica-server/commit/65ab5eaf2d2d913.

It is expected that these overrides will be subsequently removed
in favour of _config.yml settings as the Geolexica frameworks gets
more and more customizable.

* Update Jekyll-Geolexica to the latest version

* Update concepts preprocessor

Preprocessor method has been moved as of
https://github.com/geolexica/geolexica-server/pull/65.

* Render dates in subtitles where appropriate

Date on About page is possibly unnecessary, but I'm keeping it now
as it was there before.  There's an issue about this page:
https://github.com/glossarist/iev-demo-site/issues/24.