Delete unused styles from AsciiMath gem
The math.css stylesheet is provided by AsciiMath gem, and has been
added to HTMLs in 5905c79 as it is required to display HTML math
properly (HTML math is a way of presenting mathematical formulas
with regular HTML elements and CSS).  We have switched to MathML
since then, and this stylesheet is no longer used.