Specify site branding in _config.yml, and delete some template differences (#62)
* Update Geolexica-Site to 1.5.1

* Delete a dead code branch

This piece of HTML was meant to be never rendered in IEV site.
Deleting it is a better choice than enclosing with an effectively
"if false" clause (these expressions always evaluate to false for
this particular site).

* Extract a a partial template

Do it as in Jekyll-Geolexica in order to reduce another template
difference.

* Specify logos in committee section of _config.yml

The "committee" configuration entry is the intended place to set up
site branding.  Fixing that removes another unnecessary difference
between templates in this project and in Jekyll-Geolexica.