Delete unnecessary template override (#64)
Delete an HTML attribute which is now unused
(see geolexica/geolexica-server#103). Then delete
the template override as it's identical to
the original template from Jekyll-Geolexica.

* Delete unused HTML attribute

This has been removed from Jekyll-Geolexica in
https://github.com/geolexica/geolexica-server/pull/103.

* Delete unnecessary template override

As a result of the previous commit, this template override became
identical to the original template provided in Jekyll-Geolexica.
Hence, it's time to delete the override.