Eliminate another unnecessary template override (#65)
Make _layouts/concept.html override redundant and delete it.

* Delete meaningless front matter difference

There's no "registers" page in IEV, therefore having it in "nav_items"
changes nothing.  However, adding it elliminates a difference between
the original template and its override.

* Delete unnecessary template override

As a result of the previous commit, this template override became
identical to the original template provided in Jekyll-Geolexica.
Hence, it's time to delete the override.